WIP - Unable to secure connection with AWS Kinesis
Everything is statically built and linked into this app.
While it does launch on the iPad, but fails to establish secure transport with AWS Kinesis Video stream endpoint due to certificates (CURL issues).

Keeping this as a step, since the next trails will be to build using LibreSSL which is often used in Apple devices with the secure-transport option enabled.