Work revision on iPad
Got a working edition of GStreamer pipeline with AWS Kinesis Video stream working on the iPad.

Reduce video capture to a 1280 X 720 pixels at 30 frames/second as well as encoding using H.265
seems adaquate to reduce losing of frames.

Additionally, these settings keeps the network transmission rate relatively low in which this
rate is between 0.9 mbps to 1.2 mbps, but varies on the amount of change between frames.

Set logging of KVSSINK plugin to INFO, since loggin at TRACE level is in the critical path and
does affect performance somewhat.