GStreamer pipeline on iPad sink to screen, save as video as HLS, to AWS KVS
Defined a GStreamer pipeline in that captures the iPad camera (rear) to
- Display on the iPad screen
- Save locally on the iPad as a video in HLS format
- Stream to Amazon's AWS Kinesis Video Services

Missing is support for capturing the iPad audio via it's microphone. Here there is something off
with audio or audio encoding since even with the test audio source there are issues when the
pipeline transitions to a PLAYING state.